refactor: extract useEffect and useState logic into separate hook in recipeDetails
