+++ client/src/components/InsForClickedRecipeFromSearch.tsx
@@ -1,23 +1,7 @@
-import React, { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
-import { fetchRecipeDetails } from "../apiServices/apiServices";
-import { ProductDetails } from "../interfaces/general";
+import useRecipeDetails from "../hooks/useRecipeDetails";
 
 function InsForClickedRecipeFromSearch() {
-  const { recipeId } = useParams();
-  const [recipeDetails, setRecipeDetails] = useState<ProductDetails>(null);
-
-  useEffect(() => {
-    async function fetchData() {
-      try {
-        const data = await fetchRecipeDetails(recipeId);
-        setRecipeDetails(data);
-      } catch (err) {
-        console.log(err);
-      }
-    }
-    fetchData();
-  }, [recipeId]);
+  const recipeDetails = useRecipeDetails();
 
   return (
     <div className="InsClickedRecipe">
